Assignment 2
The Problem and Project
One of the main risks of a loan business is default, which means applicants borrow money from the company but fail to repay for some reason. We will use big data processing to help the company reduce this risk.
In Assignment 1, we analysed historical loan data to help the company better understand their applicants.
In Assignment 2 part A, we are going to build classification models to predict whether an applicant will default.
The Dataset
Three files from assignment 1; the following dataset is identical to assignment 1. If you still have the original files(unmodified), you don’t have to re-download them from Moodle.
1)	previous_application.csv: contains data from previous/existing applications
2)	application_data.csv: contains current application and applicant information
3)	value_dict.csv: contains a dictionary of values. Some columns in application tables are stored as integers to save storage, you need to join this table to find the meaning of those values.
Two new files(available on Moodle):
4)	loan_default.csv: contains labelled records of loan defaults. This file only has two columns: id_app and is_default.
5)	metadata.pdf(new): contains column information and data type.